Moved Scene classes into their own files
closes #59
