+++ engine-v2/src/scene_mainmenu.cpp
@@ -1,0 +1,16 @@
+#include "scene_mainmenu.h"
+#include "game.h"
+
+// Menu Scene =================================================================
+
+void MenuScene::Init() {
+
+}
+
+void MenuScene::Update() {
+
+}
+
+void MenuScene::Render() {
+
+}
