+++ engine-v2/src/scene_mainmenu.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "scene.h"
+
+struct MenuScene : Scene {
+	int test = -1;
+
+	void Init();
+	void Update();
+	void Render();
+};
